Groups units by type and open status. (#262)
If an apartment has multiple unit offerings with the same type but
different statuses, those are now displayed separately on the apartment
details page. Previously, all offerings with the same type would be
grouped together and the status of the first offering would be
incorrectly displayed for the whole group.