Clears unused inputs before generating results. (#274)
If a user enters an input that makes a follow-up page appear, enters
data into that follow-up page, and then hides the follow-up page again,
the input fields in that follow-up page are now cleared when
`computeEligibility()` is called.

As an example:
User marks "disabled" and then answers the follow-up disability service
dog question. The user goes back and un-marks "disabled". Rather than
keep the answer to the disability service dog question populated, it is
now cleared when all inputs are gathered for eligibility computation.