Returns EligResult from eligibility functions. (#302)
Instead of `true`, `false`, or `null` the eligibility functions will now
return an `EligResult` object that contains the boolean (or null) value
as well as a list of individual eligibility conditions (e.g. "Is older
than 65", "Income is below $1000 per month" and any warning/info flags
(e.g. situation is too complex to determine eligibility). These
conditions and flags can be used to render the final display of programs
for a user.

This PR does not implement any of the conditions or flags, only migrates
to the `EligResult` return value.