Changes the "call for status" badge on the housing page to "call for availability" (#611)
Before:

![343953796-dd1763c0-ead8-4f3a-b375-7f0ffa2ff492](https://github.com/theunitedeffort/theunitedeffort.org/assets/66636627/8add11ad-a626-4d53-94b5-7787e0ddf51b)
After:

![image](https://github.com/theunitedeffort/theunitedeffort.org/assets/66636627/8b9b9dec-ab35-45e5-96a1-c32de4bcea0a)

The current solution changes the status badge of a housing property to
"Call for Availability" if the returned value from Airtable is "Call for
Status", and leaves the status return value alone otherwise. In the
future, the respective field on Airtable should be changed to "Call for
Availability" as well, and the code (presently `housing.js` at line 123)
should be changed again to remove the backward compatibility
functionality. See #57.