Inverts referrals-only checkbox. (#615)
Since queries are stored in the URL parameters, we need to be careful
about backwards-compatibility for anyone who has already saved a URL for
a housing search.

This change renames the `includeReferrals` URL param to
`excludeReferrals`. That results in the following behavior for the two
possible types of legacy URLs:

- A URL with `includeReferrals=on` in the URL query string will show all
properties, including referral-only. This is because the
`includeReferrals` key is no longer valid and therefore ignored. The new
default behavior is to show referrals, so they are shown. The end result
is no change from prior behavior.
- A URL missing `includeReferrals=on` in the query string will _also_
show all properties, including referral-only. This is different from
current behavior, but opting to include _more_ results for a person's
saved URL should not be a problem. We do want to move to default-show of
referral-only properties so this supports that.

Fixes #514