Refactors CMS data to be global. (#684)
Housing searches should be as fast as possible. They run in a serverless
environment, which by default executes all global data files when
rendering. So, only the housing-related data files were global. With
this change, the CMS data file is moved back to the global dir, but
modified so that nothing happens if executed in a serverless
environment.

With the CMS data file as a directory-level data file, it was getting
called many times in succession: once for each template in that
directory. All those API calls are definitely not necessary